+++ programs/drift/src/error.rs
@@ -635,6 +635,10 @@ pub enum ErrorCode {
     InvalidSignedMsgUserOrdersResize,
     #[msg("Could not deserialize high leverage mode config")]
     CouldNotDeserializeHighLeverageModeConfig,
+    #[msg("Invalid If Rebalance Config")]
+    InvalidIfRebalanceConfig,
+    #[msg("Invalid If Rebalance Swap")]
+    InvalidIfRebalanceSwap,
 }
 
 #[macro_export]