+++ programs/drift/src/instructions/admin.rs
@@ -44,6 +44,7 @@ use crate::state::fulfillment_params::phoenix::PhoenixV1FulfillmentConfig;
 use crate::state::fulfillment_params::serum::SerumContext;
 use crate::state::fulfillment_params::serum::SerumV3FulfillmentConfig;
 use crate::state::high_leverage_mode_config::HighLeverageModeConfig;
+use crate::state::if_rebalance_config::{IfRebalanceConfig, IfRebalanceConfigParams};
 use crate::state::insurance_fund_stake::ProtocolIfSharesTransferConfig;
 use crate::state::oracle::get_sb_on_demand_price;
 use crate::state::oracle::{
@@ -4656,6 +4657,48 @@ pub fn handle_admin_deposit<'c: 'info, 'info>(
     Ok(())
 }
 
+pub fn handle_initialize_if_rebalance_config(
+    ctx: Context<InitializeIfRebalanceConfig>,
+    params: IfRebalanceConfigParams,
+) -> Result<()> {
+    let clock = Clock::get()?;
+    let now = clock.unix_timestamp;
+
+    let pubkey = ctx.accounts.if_rebalance_config.to_account_info().key;
+    let mut config = ctx.accounts.if_rebalance_config.load_init()?;
+
+    config.pubkey = *pubkey;
+    config.total_in_amount = params.total_in_amount;
+    config.current_in_amount = 0;
+    config.epoch_max_in_amount = params.epoch_max_in_amount;
+    config.epoch_duration = params.epoch_duration;
+    config.out_market_index = params.out_market_index;
+    config.in_market_index = params.in_market_index;
+    config.max_slippage_bps = params.max_slippage_bps;
+    config.swap_mode = params.swap_mode;
+    config.status = 0;
+
+    config.validate()?;
+
+    Ok(())
+}
+
+pub fn handle_update_if_rebalance_config(
+    ctx: Context<UpdateIfRebalanceConfig>,
+    params: IfRebalanceConfigParams,
+) -> Result<()> {
+    let mut config = load_mut!(ctx.accounts.if_rebalance_config)?;
+
+    config.total_in_amount = params.total_in_amount;
+    config.epoch_max_in_amount = params.epoch_max_in_amount;
+    config.epoch_duration = params.epoch_duration;
+    config.max_slippage_bps = params.max_slippage_bps;
+
+    config.validate()?;
+
+    Ok(())
+}
+
 #[derive(Accounts)]
 pub struct Initialize<'info> {
     #[account(mut)]
@@ -5428,4 +5471,37 @@ pub struct AdminDeposit<'info> {
     )]
     pub admin_token_account: Box<InterfaceAccount<'info, TokenAccount>>,
     pub token_program: Interface<'info, TokenInterface>,
+}
+
+#[derive(Accounts)]
+#[instruction(params: IfRebalanceConfigParams)]
+pub struct InitializeIfRebalanceConfig<'info> {
+    #[account(mut)]
+    pub admin: Signer<'info>,
+    #[account(
+        init,
+        seeds = [b"if_rebalance_config".as_ref(), params.in_market_index.to_le_bytes().as_ref(), params.out_market_index.to_le_bytes().as_ref()],
+        space = IfRebalanceConfig::SIZE,
+        bump,
+        payer = admin
+    )]
+    pub if_rebalance_config: AccountLoader<'info, IfRebalanceConfig>,
+    #[account(
+        has_one = admin
+    )]
+    pub state: Box<Account<'info, State>>,
+    pub rent: Sysvar<'info, Rent>,
+    pub system_program: Program<'info, System>,
+}
+
+#[derive(Accounts)]
+pub struct UpdateIfRebalanceConfig<'info> {
+    #[account(mut)]
+    pub admin: Signer<'info>,
+    #[account(mut)]
+    pub if_rebalance_config: AccountLoader<'info, IfRebalanceConfig>,
+    #[account(
+        has_one = admin
+    )]
+    pub state: Box<Account<'info, State>>,
 }