+++ sdk/src/adminClient.ts
@@ -14,6 +14,7 @@ import {
 	ContractTier,
 	AssetTier,
 	SpotFulfillmentConfigStatus,
+	IfRebalanceConfigParams,
 } from './types';
 import { DEFAULT_MARKET_NAME, encodeName } from './userName';
 import { BN } from '@coral-xyz/anchor';
@@ -37,6 +38,7 @@ import {
 	getProtectedMakerModeConfigPublicKey,
 	getFuelOverflowAccountPublicKey,
 	getTokenProgramForSpotMarket,
+	getIfRebalanceConfigPublicKey,
 } from './addresses/pda';
 import { squareRootBN } from './math/utils';
 import { TOKEN_PROGRAM_ID } from '@solana/spl-token';
@@ -3618,6 +3620,65 @@ export class AdminClient extends DriftClient {
 		);
 	}
 
+	public async transferProtocolIfSharesToRevenuePool(
+		outMarketIndex: number,
+		inMarketIndex: number,
+		amount: BN
+	): Promise<TransactionSignature> {
+		const transferProtocolIfSharesToRevenuePoolIx =
+			await this.getTransferProtocolIfSharesToRevenuePoolIx(
+				outMarketIndex,
+				inMarketIndex,
+				amount
+			);
+
+		const tx = await this.buildTransaction(
+			transferProtocolIfSharesToRevenuePoolIx
+		);
+
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+
+		return txSig;
+	}
+
+	public async getTransferProtocolIfSharesToRevenuePoolIx(
+		outMarketIndex: number,
+		inMarketIndex: number,
+		amount: BN
+	): Promise<TransactionInstruction> {
+		const remainingAccounts = await this.getRemainingAccounts({
+			userAccounts: [],
+			writableSpotMarketIndexes: [outMarketIndex],
+		});
+
+		return await this.program.instruction.transferProtocolIfSharesToRevenuePool(
+			outMarketIndex,
+			amount,
+			{
+				accounts: {
+					authority: this.wallet.publicKey,
+					state: await this.getStatePublicKey(),
+					insuranceFundVault: await getInsuranceFundVaultPublicKey(
+						this.program.programId,
+						outMarketIndex
+					),
+					spotMarketVault: await getSpotMarketVaultPublicKey(
+						this.program.programId,
+						outMarketIndex
+					),
+					ifRebalanceConfig: await getIfRebalanceConfigPublicKey(
+						this.program.programId,
+						inMarketIndex,
+						outMarketIndex
+					),
+					tokenProgram: TOKEN_PROGRAM_ID,
+					driftSigner: this.getStateAccount().signer,
+				},
+				remainingAccounts,
+			}
+		);
+	}
+
 	public async initializePrelaunchOracle(
 		perpMarketIndex: number,
 		price?: BN,
@@ -3982,6 +4043,38 @@ export class AdminClient extends DriftClient {
 				},
 			}
 		);
+	}
+
+	public async initializeIfRebalanceConfig(
+		params: IfRebalanceConfigParams
+	): Promise<TransactionSignature> {
+		const initializeIfRebalanceConfigIx =
+			await this.getInitializeIfRebalanceConfigIx(params);
+
+		const tx = await this.buildTransaction(initializeIfRebalanceConfigIx);
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+
+		return txSig;
+	}
+
+	public async getInitializeIfRebalanceConfigIx(
+		params: IfRebalanceConfigParams
+	): Promise<TransactionInstruction> {
+		return await this.program.instruction.initializeIfRebalanceConfig(params, {
+			accounts: {
+				admin: this.isSubscribed
+					? this.getStateAccount().admin
+					: this.wallet.publicKey,
+				state: await this.getStatePublicKey(),
+				ifRebalanceConfig: await getIfRebalanceConfigPublicKey(
+					this.program.programId,
+					params.inMarketIndex,
+					params.outMarketIndex
+				),
+				rent: SYSVAR_RENT_PUBKEY,
+				systemProgram: anchor.web3.SystemProgram.programId,
+			},
+		});
 	}
 
 	public async initUserFuel(